+++ packages/query-sync-storage-persister/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @tanstack/query-sync-storage-persister
 
+## 5.90.21
+
+### Patch Changes
+
+- Updated dependencies [[`53fc74e`](https://github.com/TanStack/query/commit/53fc74ebb16730bd3317f039a69c6821386bae93)]:
+  - @tanstack/query-core@5.90.19
+  - @tanstack/query-persist-client-core@5.91.18
+
+## 5.90.20
+
+### Patch Changes
+
+- Updated dependencies [[`dea1614`](https://github.com/TanStack/query/commit/dea1614aaad5c572cf43cea54b64ac09dc4d5b41)]:
+  - @tanstack/query-core@5.90.18
+  - @tanstack/query-persist-client-core@5.91.17
+
 ## 5.90.19
 
 ### Patch Changes